Created new 'scenes' directory
Added:
- new directory to store phaser js scenes
- new file 'game' that will contain the code for the main scene

The phaser scene was moved to its own directory to free up app.js.

Store.js will contain global variables that I can access from within the
program. The first being an index of my game scenes.

The scenes will be pushed to this array after it calls the next scene to
make sure they are in the correct order.

* Next Commit
The webpack config needs to involve phaser assets in its build process.
Going to try experimenting with making a bundle for phasers code and
experimenting with the common chunks plugin.