Complete user auth ui events with Toastr implementation
Toastr is a javascript library for non-blocking notifications. It will
be used to display unique ui messages.

toastr.js contains the configuration for the Toastr Class. The parameters
for the toastr instantiation are defined inside javascript objects.
The ui events inside `ui.js` traverse these objects to call the
appropriate toastr notification.